Mix directly into ApplicationController as some gems may have already loaded.
Some gems (like resource_controller) may load the ApplicationController during
gem loading (i.e. prior to gem init being run). This mean that if we inject
on functionality on ActionController::Base some things will not carry through
to the child already subclassed (like before filters). So we need to inject
directly on the ApplicationController.

But to avoid creating the problem that resource_controller has created for
us we inject it on the to_prepare callback and have some protection against
installing hooks more than  once.