Major refactoring and full tests.
There was a bit of cruft that was put in over time due to certain needs without
really thinking about the best way to go about it. This gets rid of all that
cruft.

It also introduces tests with my new test_engine gem. Full tests is in place
(unit, functional and integration). Obviously this was a bit backwards (would
have been nice to test first) but since Rails didn't really support testing
engines we went about it backwards.

Although this engine has obviously be tested well it is not integrated into
any production app yet but it will soon. So expect another patch or two
to come down the pipe.