Directly monkey-patch instead of trying to override via inheritance.
Using inheritance to override meant we required the use of ActiveRecord. This
way we can intergrate with anything AuthLogic can integrate with (which by
itself is only ActiveRecord but with mm_authlogic can also be MongoMapper).

Monkey-patching is obviously fragile but not really any more so than the
inheritance setup.