More Rails 4 compat
Marking Session as unloadable was causing a circular dependency in development
mode anytime a model changed. Taking that out and just letting Rails handle
everything seemed to resolve it.

Based on the old comment this is just an optimization so it should not effect
older versions of Rails. But the comment could be wrong so we will see.