Update Gemfile to test against latest and make tests Rails 4 friendly.
The gem should still work on Rails 3. But the tests were not running under
Rails 4 so switched them to focus on that. They may also still run under
Rails 3.

There are multiple deprecation warnings due to Authlogic. The development
version of Authlogic doesn't have those so use that in your app if you wan
to get rid of them.