Compat with newer Rails.
The `before_filter` is going away so switching to `before_action`.
Also we can no longer just `include` on ActionController loading as
it includes on a lower-level class that doesn't have helpers.

I updated the dependencies in an attempt to run the tests but I could
not run them. The `test_engine` gem I am using no longer seems to work
with newer dependencies. Given the changes to this gem are only minor
and that we are no longer maintaining this gem I am just going to
release without test.

Also noted the lack of maintenance on this gem to warn people away
from uing it. Need to do the same for `remindme` and the `test_engine`
gem.